Update usbd_custom_hid_if.c
Fix for incorrect values. Looks like the two sections for slider inputs were copy/pasted and both ended up being identified with the same value 0x36 so changed the second to 0x37. Also looks like the pointer values on lines 140 and 151 were also mistakenly set to 0x36 in the process. Corrected to match usage on lines 114 and 127

Could be root cause of improper HID reporting on linux systems